Split docker test fixtures out so they can come up and down in parallel.
