+++ mlos_bench/mlos_bench/tests/storage/conftest.py
@@ -4,6 +4,7 @@
 #
 """Export test fixtures for mlos_bench storage."""
 
+import mlos_bench.tests.docker_fixtures_util as docker_fixtures_util
 import mlos_bench.tests.storage.sql.fixtures as sql_storage_fixtures
 
 # NOTE: For future storage implementation additions, we can refactor this to use
@@ -11,6 +12,14 @@ import mlos_bench.tests.storage.sql.fixtures as sql_storage_fixtures
 # same.
 
 # Expose some of those as local names so they can be picked up as fixtures by pytest.
+docker_compose_file = sql_storage_fixtures.docker_compose_file
+docker_compose_project_name = sql_storage_fixtures.docker_compose_project_name
+
+docker_setup = docker_fixtures_util.docker_setup
+docker_services_lock = docker_fixtures_util.docker_services_lock
+docker_setup_teardown_lock = docker_fixtures_util.docker_setup_teardown_lock
+locked_docker_services = docker_fixtures_util.locked_docker_services
+
 mysql_storage_info = sql_storage_fixtures.mysql_storage_info
 mysql_storage = sql_storage_fixtures.mysql_storage
 postgres_storage_info = sql_storage_fixtures.postgres_storage_info
